Add hamlib support for push to talk
This removes libctb in favour of hamlib. It has been tested with 1.2.15.3 of
hamlib with a IC-7200.

The rig is configured using it's own dialog Tools -> PTT Config.  The only
options available are the rig model and serial port, the rest of the options
are obtained from the rig definition in hamlib.

FreeDV will open the configured serial port on startup. When it cannot find
the radio, an error dialog will be displayed. Spacebar can be used to toggle
TX/RX.


git-svn-id: https://svn.code.sourceforge.net/p/freetel/code@1283 01035d8c-6547-0410-b346-abe4f91aad63