sm1000_main: Re-calibrate SysTick timer
Not sure why, but SysTick was being calibrated with a period equal to
the core frequency divided by 168000.

Apparently this was supposed to give millisecond ticks.  SystemCoreClock
gives the frequency of the core clock, and if you supply this value to
SysTick_Config, you get a 1 second tick timer.

Dividing this by 1000 seems to give much better timing results.  The
time-out timer almost works now.

git-svn-id: https://svn.code.sourceforge.net/p/freetel/code@2470 01035d8c-6547-0410-b346-abe4f91aad63